[pre-reease] unit test example for data endpoint
