+++ package.json
@@ -24,6 +24,7 @@
     "crypto-browserify": "^3.12.0",
     "jest": "^29.7.0",
     "jest-mock-server": "^0.1.0",
+    "jest-websocket-mock": "^2.5.0",
     "node-polyfill-webpack-plugin": "^3.0.0",
     "openurl": "^1.1.1",
     "path-browserify": "^1.0.1",
