+++ src/eyepop/data/data_options.ts
@@ -5,5 +5,6 @@ export interface DataOptions extends Options {
     /**
      * Defaults to process.env['EYEPOP_ACCOUNT_ID'].
      */
-    accountId?: string | undefined;
+    accountId?: string | undefined
+    disableWs?: boolean | undefined
 }
