+++ tests/eyepop/worker/endpoint.connect.test.ts
@@ -1,4 +1,4 @@
-import {EyePop, Session} from '../../src/eyepop'
+import {EyePop, Session} from '../../../src/eyepop'
 
 import {MockServer} from 'jest-mock-server'
 import {describe, expect, test} from '@jest/globals'
