+++ tests/eyepop/worker/endpoint.load.test.ts
@@ -1,10 +1,10 @@
-import {EyePop} from '../../src/eyepop'
+import {EyePop} from '../../../src/eyepop'
 
 import {MockServer} from 'jest-mock-server'
 import {describe, expect, test} from '@jest/globals'
 import {v4 as uuidv4} from 'uuid'
 
-function prepMockServer(server: MockServer, test_pop_id: string, test_pipeline_id_1: string, test_pipeline_id_2: string) {
+function prepMockServer(server: MockServer, test_pop_id: string, test_pipeline_id: string) {
     const test_access_token = uuidv4()
     const token_valid_time = 1000 * 1000
 
@@ -25,37 +25,23 @@ function prepMockServer(server: MockServer, test_pop_id: string, test_pipeline_i
         .mockImplementationOnce((ctx) => {
             ctx.status = 200
             ctx.response.headers['content-type'] = 'application/json'
-            ctx.body = JSON.stringify({base_url: `${server.getURL()}worker/`, pipeline_id: test_pipeline_id_1})
-        })
-        .mockImplementationOnce((ctx) => {
-            ctx.status = 200
-            ctx.response.headers['content-type'] = 'application/json'
-            ctx.body = JSON.stringify({base_url: `${server.getURL()}worker/`, pipeline_id: test_pipeline_id_2})
+            ctx.body = JSON.stringify({base_url: `${server.getURL()}worker/`, pipeline_id: test_pipeline_id})
         })
 
-    const getPipelineRoute1 = server
-        .get(`/worker/pipelines/${test_pipeline_id_1}`)
+    const getPipelineRoute = server
+        .get(`/worker/pipelines/${test_pipeline_id}`)
         .mockImplementationOnce((ctx) => {
             ctx.status = 200
             ctx.response.headers['content-type'] = 'application/json'
             ctx.body = JSON.stringify({
-                id: test_pipeline_id_1
-            })
-        })
-    const getPipelineRoute2 = server
-        .get(`/worker/pipelines/${test_pipeline_id_2}`)
-        .mockImplementationOnce((ctx) => {
-            ctx.status = 200
-            ctx.response.headers['content-type'] = 'application/json'
-            ctx.body = JSON.stringify({
-                id: test_pipeline_id_2
+                id: test_pipeline_id
             })
         })
 
-    return {authenticationRoute, popConfigRoute ,getPipelineRoute1, getPipelineRoute2};
+    return {authenticationRoute, popConfigRoute ,getPipelineRoute};
 }
 
-describe('EyePopSdk endpoint reconnect', () => {
+describe('EyePopSdk endpoint module loadFrom', () => {
     const server = new MockServer()
 
     const test_secret_key = uuidv4()
@@ -64,23 +50,16 @@ describe('EyePopSdk endpoint reconnect', () => {
     afterAll(() => server.stop())
     beforeEach(() => server.reset())
 
-    test('EyePopSdk loadFrom reload config on 404', async () => {
+    test('EyePopSdk loadFrom', async () => {
         const fake_timestamp = Date.now()
         const test_pop_id = uuidv4()
-        const test_pipeline_id_1 = uuidv4()
-        const test_pipeline_id_2 = uuidv4()
+        const test_pipeline_id = uuidv4()
         const location = 'http://invalid.example'
-        const {authenticationRoute, popConfigRoute} =
-            prepMockServer(server, test_pop_id, test_pipeline_id_1, test_pipeline_id_2)
+        const {authenticationRoute, popConfigRoute} = prepMockServer(server, test_pop_id, test_pipeline_id)
 
-        const loadFromRoute1 = server
-        .patch(`/worker/pipelines/${test_pipeline_id_1}/source`)
-        .mockImplementationOnce(async (ctx) => {
-            ctx.status = 404
-        })
-        const loadFromRoute2 = server
-        .patch(`/worker/pipelines/${test_pipeline_id_2}/source`)
-        .mockImplementationOnce(async (ctx) => {
+        const loadFromRoute = server
+        .patch(`/worker/pipelines/${test_pipeline_id}/source`)
+        .mockImplementation(async (ctx) => {
             expect(ctx.headers['authorization']).toBeDefined()
             expect(ctx.request.query['mode']).toBe('queue')
             expect(ctx.request.query['processing']).toBe('sync')
@@ -112,8 +91,7 @@ describe('EyePopSdk endpoint reconnect', () => {
                 count++
                 expect(prediction.timestamp).toBe(fake_timestamp)
             }
-            expect(loadFromRoute1).toHaveBeenCalledTimes(1)
-            expect(loadFromRoute2).toHaveBeenCalledTimes(1)
+            expect(loadFromRoute).toHaveBeenCalledTimes(1)
             expect(count).toBe(1)
         } finally {
             await endpoint.disconnect()
