+++ tests/eyepop/worker/endpoint.upload.test.ts
@@ -1,8 +1,9 @@
-import {EyePop} from '../../src/eyepop'
+import {EyePop} from '../../../src/eyepop'
 
 import {MockServer} from 'jest-mock-server'
 import {describe, expect, test} from '@jest/globals'
 import {v4 as uuidv4} from 'uuid'
+import * as fs from "fs"
 
 function prepMockServer(server: MockServer, test_pop_id: string, test_pipeline_id: string) {
     const test_access_token = uuidv4()
@@ -14,9 +15,9 @@ function prepMockServer(server: MockServer, test_pop_id: string, test_pipeline_i
             ctx.status = 200
             ctx.response.headers['content-type'] = 'application/json'
             ctx.body = JSON.stringify({
-                'access_token': test_access_token,
-                'expires_in': token_valid_time,
-                'token_type': 'Bearer'
+                access_token: test_access_token,
+                expires_in: token_valid_time,
+                token_type: 'Bearer'
             })
         })
 
@@ -41,7 +42,7 @@ function prepMockServer(server: MockServer, test_pop_id: string, test_pipeline_i
     return {authenticationRoute, popConfigRoute ,getPipelineRoute};
 }
 
-describe('EyePopSdk endpoint module loadFrom', () => {
+describe('EyePopSdk endpoint module upload', () => {
     const server = new MockServer()
 
     const test_secret_key = uuidv4()
@@ -50,24 +51,23 @@ describe('EyePopSdk endpoint module loadFrom', () => {
     afterAll(() => server.stop())
     beforeEach(() => server.reset())
 
-    test('EyePopSdk loadFrom', async () => {
+    test('EyePopSdk upload', async () => {
         const fake_timestamp = Date.now()
         const test_pop_id = uuidv4()
         const test_pipeline_id = uuidv4()
-        const location = 'http://invalid.example'
+        const image_path = './tests/test.jpg'
         const {authenticationRoute, popConfigRoute} = prepMockServer(server, test_pop_id, test_pipeline_id)
 
-        const loadFromRoute = server
-        .patch(`/worker/pipelines/${test_pipeline_id}/source`)
+        const uploadRoute = server
+        .post(`/worker/pipelines/${test_pipeline_id}/source`)
         .mockImplementation(async (ctx) => {
             expect(ctx.headers['authorization']).toBeDefined()
             expect(ctx.request.query['mode']).toBe('queue')
             expect(ctx.request.query['processing']).toBe('sync')
-            expect(ctx.request.body).toBeDefined()
-            // @ts-ignore
-            expect(ctx.request.body['sourceType']).toBe('URL')
-            // @ts-ignore
-            expect(ctx.request.body['url']).toBe(location)
+            let bodyContent = await ctx.request.req.toArray()
+            // post is gzip'ed and not sure how to use this KOA api to retrieve the decompressed content
+            expect(bodyContent.length).toBeGreaterThan(0)
+            expect(bodyContent.length).toBeLessThanOrEqual(fs.statSync(image_path).size)
             ctx.status = 200
             ctx.response.headers['content-type'] = 'application/json'
             ctx.body = JSON.stringify({timestamp: fake_timestamp})
@@ -84,15 +84,39 @@ describe('EyePopSdk endpoint module loadFrom', () => {
             await endpoint.connect()
             expect(authenticationRoute).toHaveBeenCalledTimes(1)
             expect(popConfigRoute).toHaveBeenCalledTimes(1)
-            let job = await endpoint.process({url: location})
+            let job = await endpoint.process({path: image_path})
             expect(job).toBeDefined()
             let count = 0
-            for await (let prediction of job) {
+            for await (let prediction of await job) {
                 count++
                 expect(prediction.timestamp).toBe(fake_timestamp)
             }
-            expect(loadFromRoute).toHaveBeenCalledTimes(1)
+            expect(uploadRoute).toHaveBeenCalledTimes(1)
             expect(count).toBe(1)
+        } finally {
+            await endpoint.disconnect()
+        }
+    })
+
+      test('EyePopSdk upload file not found', async () => {
+        const fake_timestamp = Date.now()
+        const test_pop_id = uuidv4()
+        const test_pipeline_id = uuidv4()
+        const image_path = './tests/does_not_exist.dummy'
+        const {authenticationRoute, popConfigRoute} = prepMockServer(server, test_pop_id, test_pipeline_id)
+
+        const endpoint = EyePop.workerEndpoint({
+            eyepopUrl: server.getURL().toString(),
+            auth: {secretKey: test_secret_key},
+            popId: test_pop_id,
+            stopJobs: false
+        })
+        expect(endpoint).toBeDefined()
+        try {
+            await endpoint.connect()
+            expect(authenticationRoute).toHaveBeenCalledTimes(1)
+            expect(popConfigRoute).toHaveBeenCalledTimes(1)
+            expect(endpoint.process({path: image_path})).rejects.toBeDefined()
         } finally {
             await endpoint.disconnect()
         }
